add methods + new structure
