+++ models/TaskList.py
@@ -1,15 +1,3 @@
-class Task:
-    def __init__(self, name, description):
-        self.name = name
-        self.description = description
-        self.completed = False
-
-    def mark_completed(self):
-        self.completed = True
-
-    def __str__(self):
-        return f"{self.name} - {self.description} (Completed: {self.completed})"
-
 
 class TaskList:
     def __init__(self):
