+++ project_big_data/pyproject.toml
@@ -7,6 +7,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.9"
+datetime = "^5.2"
 
 
 [build-system]
